iface: rename libsidiface_usid to libsidiface
Also change usid.[ch] to iface.[ch]. No functional changes.
