+++ src/resource/Makefile.am
@@ -44,7 +44,7 @@ libsidresource_la_LDFLAGS = -version-info 0:0:0
 
 libsidresource_la_LIBADD = $(top_builddir)/src/base/libsidbase.la \
 			   $(top_builddir)/src/iface/libsidiface_servicelink.la \
-			   $(top_builddir)/src/iface/libsidiface_usid.la \
+			   $(top_builddir)/src/iface/libsidiface.la \
 			   $(top_builddir)/src/log/libsidlog.la \
 			   $(SYSTEMD_LIBS) \
 			   $(UDEV_LIBS) \
