+++ src/tools/usid/Makefile.am
@@ -24,6 +24,6 @@ udev_PROGRAMS = usid
 usid_SOURCES = usid.c
 
 usid_LDADD = $(top_builddir)/src/base/libsidbase.la \
-	     $(top_builddir)/src/iface/libsidiface_usid.la \
+	     $(top_builddir)/src/iface/libsidiface.la \
 	     $(top_builddir)/src/log/libsidlog.la
